+++ tests/unit/Auth/loginCommand.spec.ts
@@ -1,18 +1,39 @@
 import * as should from 'should';
-import { getIsAuth0Enabled } from '../../../src/Auth/loginCommand';
+import { getIsAuth0OrLegacyEnabled } from '../../../src/Auth/loginCommand';
 
 describe('Auth.loginCommand', () => {
-	describe('getIsAuth0Enabled', () => {
+	describe('getIsAuth0OrLegacyEnabled', () => {
 		it('should return true if --auth0-enabled is present', () => {
-			const isAuth0Enabled = getIsAuth0Enabled({ _unknown: ['--auth0-enabled'], command: ['login'] });
+			const { isAuth0Enabled } = getIsAuth0OrLegacyEnabled({ _unknown: ['--auth0-enabled'], command: ['login'] });
 
 			should(isAuth0Enabled).be.true();
 		});
 
+		it('should return true if --legacy-enabled is present', () => {
+			const { isLegacyEnabled } = getIsAuth0OrLegacyEnabled({ _unknown: ['--legacy-enabled'], command: ['login'] });
+
+			should(isLegacyEnabled).be.true();
+		});
+
 		it('should return false if --auth0-enabled is NOT present', () => {
-			const isAuth0Enabled = getIsAuth0Enabled({});
+			const { isAuth0Enabled } = getIsAuth0OrLegacyEnabled({});
 
 			should(isAuth0Enabled).be.false();
+		});
+
+		it('should return false if --legacy-enabled is NOT present', () => {
+			const { isLegacyEnabled } = getIsAuth0OrLegacyEnabled({});
+
+			should(isLegacyEnabled).be.false();
+		});
+
+		it('should fail if both --auth0-enabled and --legacy-enabled are present', () => {
+			should(() =>
+				getIsAuth0OrLegacyEnabled({
+					_unknown: ['--auth0-enabled', '--legacy-enabled'],
+					command: ['login'],
+				}),
+			).throwError('Only one override from auth0 and legacy authentication options can be active at the moment.');
 		});
 	});
 });